ref: rename `logger` to `log` pkg
