+++ log/console_writer_test.go
@@ -1,4 +1,4 @@
-package logger
+package log
 
 import (
 	"os"
